# docs/swagger.yaml
basePath: /
definitions:
  database.ActionStore:
    properties:
      actionId:
        type: integer
      address:
        type: string
      name:
        type: string
      point:
        type: integer
      time:
        type: string
    type: object
...
  types.ListProjectsRes:
    properties:
      projects:
        items:
          $ref: '#/definitions/types.ProjectInfo'
        type: array
    type: object
...
  types.PointHistoryRes:
    properties:
      history:
        items:
          $ref: '#/definitions/database.ActionStore'
        type: array
      length:
        type: integer
    type: object
  types.PointInfoRes:
    properties:
      charging:
        type: boolean
      chargingCount:
        type: integer
      godataCount:
        type: integer
      godataSpace:
        type: integer
      points:
        type: integer
    type: object
  types.ProjectInfo:
    properties:
      end:
        type: string
      name:
        type: string
      projectID:
        type: integer
      start:
        type: string
    type: object
  types.RankInfo:
    properties:
      address:
        type: string
      points:
        type: integer
      rank:
        type: integer
      scores:
        type: integer
    type: object
  types.RankRes:
    properties:
      length:
        type: integer
      rankInfo:
        items:
          $ref: '#/definitions/types.RankInfo'
        type: array
    type: object
...
  types.UserInfoRes:
    properties:
      address:
        type: string
      inviteCode:
        type: string
      invitedCode:
        type: string
      points:
        type: integer
      referrals:
        type: integer
      space:
        type: integer
      storage:
        type: integer
    type: object
host: test-xs-api.memolabs.net
info:
  contact: {}
  description: This is the API document for xspace server.
  title: Xspace-Server API
  version: "1.0"
paths:
...
  /v1/point/add:
    post:
      consumes:
      - application/json
      description: Users can earn point with finish action
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: action id (101 for follow twitter, 102 for follow discord, 103
          for follow telegram)
        in: body
        name: actionId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserInfoRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - Point
  /v1/point/charge:
    post:
      consumes:
      - application/json
      description: Users can charge once every 6 hours
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserInfoRes'
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - Point
  /v1/point/history:
    get:
      consumes:
      - application/json
      description: Get the history of the point info by address
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Pages
        in: query
        name: page
        required: true
        type: string
      - description: The amount of data displayed on each page
        in: query
        name: size
        required: true
        type: string
      - description: The action id
        in: query
        name: actionID
        type: string
      - description: Order rules (date_asc for sorting by creation time from smallest
          to largest, date_desc for sorting by creation time from largest to smallest,
          dealut is date_desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PointHistoryRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - Point
  /v1/point/info:
    get:
      consumes:
      - application/json
      description: Get the user's point info by address
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PointInfoRes'
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - User
  /v1/point/invite:
    post:
      consumes:
      - application/json
      description: Get the history of the point info by address
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: The invite code from other user
        in: body
        name: code
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserInfoRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - Point
  /v1/project/list:
    get:
      consumes:
      - application/json
      description: List all projects with Xspace
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListProjectsRes'
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - Rank
  /v1/project/rank:
    get:
      consumes:
      - application/json
      description: Get the ranking of cooperative projects
      parameters:
      - description: cooperative project id
        in: query
        name: id
        required: true
        type: string
      - description: Pages
        in: query
        name: page
        required: true
        type: string
      - description: The amount of data displayed on each page
        in: query
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.RankRes'
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - Rank
...
  /v1/user/info:
    get:
      consumes:
      - application/json
      description: Get the user basic info by address
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserInfoRes'
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - User
swagger: "2.0"